# perf(bloom-filter): remove while loop for setbit & getbit