## Description

Constant time `setbit` and `getbit` in the bloom filter - was hitting performance issues when using a large bloom filter

## Notes & open questions

Benchmarks below

```
Loop 8bits x 165,949,213 ops/sec ±3.93% (88 runs sampled)
Math Ops 8bits x 219,049,980 ops/sec ±5.85% (79 runs sampled)

Loop 64bits x 122,952,916 ops/sec ±3.76% (85 runs sampled)
Math Ops 64bits x 204,311,730 ops/sec ±6.76% (73 runs sampled)

Loop 512bits x 35,816,953 ops/sec ±1.12% (92 runs sampled)
Math Ops 512bits x 210,229,964 ops/sec ±6.10% (76 runs sampled)

Loop 4096bits x 4,882,100 ops/sec ±0.26% (98 runs sampled)
Math Ops 4096bits x 217,126,816 ops/sec ±6.68% (79 runs sampled)

Loop 32768bits x 579,858 ops/sec ±0.52% (98 runs sampled)
Math Ops 32768bits x 220,399,489 ops/sec ±5.58% (80 runs sampled)

Loop 262144bits x 61,032 ops/sec ±0.16% (98 runs sampled)
Math Ops 262144bits x 201,626,561 ops/sec ±6.81% (72 runs sampled)

Loop 262145bits x 61,022 ops/sec ±0.17% (103 runs sampled)
Math Ops 262145bits x 217,429,547 ops/sec ±5.77% (78 runs sampled)

Loop 2097152bits x 7,649 ops/sec ±0.15% (102 runs sampled)
Math Ops 2097152bits x 221,095,958 ops/sec ±6.40% (78 runs sampled)
```

## Change checklist

- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation if necessary (this includes comments as well)
- [x] I have added tests that prove my fix is effective or that my feature works