@achingbrain 2024-09-09
This looks good, thanks for opening it.

Out of interest, where are you using `BloomFilter`?

I can only see it used once in the libp2p codebase (for filtering circuit relay servers) - in general I think you'd want to use `ScalableCuckooFilter` instead as it can grow beyond the initially configured size.