@dozyio 2024-09-09
> This looks good, thanks for opening it.
> 
> Out of interest, where are you using `BloomFilter`?
> 
> I can only see it used once in the libp2p codebase (for filtering circuit relay servers) - in general I think you'd want to use `ScalableCuckooFilter` instead as it can grow beyond the initially configured size.

Hey Alex - I'm using it in https://github.com/dozyio/js-ds-crdt to check if keys are deleted from a set - was going to use [bloom-filters](https://github.com/Callidon/bloom-filters) but saw your issue https://github.com/Callidon/bloom-filters/issues/70 - the one in utils should be fine once this pr is applied