@achingbrain 2024-09-10
Fair enough.  This implementation might get removed in the future since we could just use a cuckoo filter for filtering relay servers but hopefully the `bloom-filters` module will ship it's v4 at some point soon.