@folkvir 2025-01-08
@achingbrain ✋ What is the feature you are looking for with the v4 of `bloom-filters` exactly?